remove duplicate ts file
